LFN auto completion for rmdir.

+++ DataManagementSystem/Client/FileCatalogClientCLI.py
@@ -507,7 +507,26 @@ File Catalog Client $Revision: 1.17 $Date:
     """  
     # Not yet really implemented yet
     argss = args.split()
+    if len(argss) != 1:
+      print self.do_rmdir.__doc__
+      return
     self.removeDirectory(argss)  
+
+  def complete_rmdir(self, text, line, begidx, endidx):
+    result = []
+    args = line.split()
+
+    # the first argument -- LFN.
+    if (1<=len(args)<=2):
+      # If last char is ' ',
+      # this can be a new parameter.
+      if (len(args) == 1) or (len(args)==2 and (not line.endswith(' '))):
+        cur_path = ""
+        if (len(args) == 2):
+          cur_path = args[1]
+        result = self.lfn_dc.parse_text_line(text, cur_path, self.cwd)
+
+    return result
           
   def removeReplica(self,args):
     """ Remove replica from the catalog
